requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.1.3
et-xmlfile==1.1.0
Flask==2.1.2
gcloud==0.18.3
googleapis-common-protos==1.56.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.1
numpy==1.21.3
oauth2client==4.1.3
openai==0.11.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
protobuf==4.21.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.14.1
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jwt==2.0.1
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==4.2.0
urllib3==1.26.7
Werkzeug==2.1.2
zipp==3.8.0
